CLIP-Gradio-example/requirements.txt
torch
torchvision
transformers
pandas
scikit-learn
Pillow
gradio
